pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.0"
description = ""
authors = ["gokamoda <gokamoda@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10.5"
requests = "^2.31.0"
hydra-core = "^1.3.2"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pylint = "^3.0.3"


[tool.pylint]
source-roots = ["src"]

[tool.pylint.'MESSAGES CONTROL']
disable = """
    no-value-for-parameter,
    missing-module-docstring
"""

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

